修改为size_t

+++ crypto/iso10126padding.cpp
@@ -33,7 +33,7 @@ void crypto::padding::ISO10126Padding::Pad(vector<byte>& in_out) const
 	out.insert(out.end(), padding.begin(), padding.end());
 }
 
-int crypto::padding::ISO10126Padding::Unpad(vector<byte>& in_out) const
+size_t crypto::padding::ISO10126Padding::Unpad(vector<byte>& in_out) const
 {
 	vector<byte>& in = in_out;
 	vector<byte>& out = in_out;
@@ -55,7 +55,7 @@ int crypto::padding::ISO10126Padding::Unpad(vector<byte>& in_out) const
 	return start;
 }
 
-int crypto::padding::ISO10126Padding::GetPadLength(const size_t len) const
+size_t crypto::padding::ISO10126Padding::GetPadLength(const size_t len) const
 {
 	return block_size_ - len % block_size_;
 }
