+++ crypto/nopadding.cpp
@@ -31,7 +31,7 @@ void crypto::padding::NoPadding::Pad(vector<byte>& in_out) const
 	}
 }
 
-int crypto::padding::NoPadding::Unpad(vector<byte>& in_out) const
+size_t crypto::padding::NoPadding::Unpad(vector<byte>& in_out) const
 {
 	vector<byte>& in = in_out;
 	vector<byte>& out = in_out;
@@ -62,7 +62,7 @@ int crypto::padding::NoPadding::Unpad(vector<byte>& in_out) const
 	return start;
 }
 
-int crypto::padding::NoPadding::GetPadLength(size_t len) const
+size_t crypto::padding::NoPadding::GetPadLength(size_t len) const
 {
 	return (len / block_size_ + 1) * block_size_ - len;
 }
