+++ crypto/oaepping.cpp
@@ -71,7 +71,7 @@ void crypto::padding::OAEPwithHashandMGF1Padding::Pad(vector<byte>& in_out) cons
 	out = padded;
 }
 
-int crypto::padding::OAEPwithHashandMGF1Padding::Unpad(vector<byte>& in_out) const
+size_t crypto::padding::OAEPwithHashandMGF1Padding::Unpad(vector<byte>& in_out) const
 {
 	vector<byte>& in = in_out;
 	vector<byte>& out = in_out;
@@ -98,10 +98,10 @@ int crypto::padding::OAEPwithHashandMGF1Padding::Unpad(vector<byte>& in_out) con
 	return result.messageLength;
 }
 
-int crypto::padding::OAEPwithHashandMGF1Padding::GetPadLength(const size_t len) const
+size_t crypto::padding::OAEPwithHashandMGF1Padding::GetPadLength(const size_t len) const
 {
 	CryptoPP::OAEP_Base* oaep = GetOAEPFunction();
-	const int padding_size = oaep->MaxUnpaddedLength(block_size_ * 8) - len;
+	const size_t padding_size = oaep->MaxUnpaddedLength(block_size_ * 8) - len;
 	delete oaep;
 	return padding_size;
 }
