+++ crypto/padding.h
@@ -38,12 +38,12 @@ public:
 	 * \return the index where the padding starts, or -1 if the input is
 	 * not properly padded
 	 */
-	virtual int Unpad(vector<byte>& in) const = 0;
+	virtual size_t Unpad(vector<byte>& in) const = 0;
 
 	/**
 	 * \brief Determines how long the padding will be for a given input length.
 	 * \param len the length of the data to pad
 	 * \return the length of the padding
 	 */
-	virtual int GetPadLength(size_t len) const = 0;
+	virtual size_t GetPadLength(size_t len) const = 0;
 };
