+++ crypto/pkcs5padding.cpp
@@ -24,7 +24,7 @@ void crypto::padding::PKCS5Padding::Pad(vector<byte>& in_out) const
 	out.insert(out.end(), len, padding_octet);
 }
 
-int crypto::padding::PKCS5Padding::Unpad(vector<byte>& in_out) const
+size_t crypto::padding::PKCS5Padding::Unpad(vector<byte>& in_out) const
 {
 	vector<byte>& in = in_out;
 	vector<byte>& out = in_out;
@@ -52,7 +52,7 @@ int crypto::padding::PKCS5Padding::Unpad(vector<byte>& in_out) const
 	return start;
 }
 
-int crypto::padding::PKCS5Padding::GetPadLength(const size_t len) const
+size_t crypto::padding::PKCS5Padding::GetPadLength(const size_t len) const
 {
 	return block_size_ - len % block_size_;
 }
