+++ crypto/x932padding.cpp
@@ -24,7 +24,7 @@ void crypto::padding::X932Padding::Pad(vector<byte>& in_out) const
 	out.insert(out.end(), 1, static_cast<byte>(len & 0xFF));
 }
 
-int crypto::padding::X932Padding::Unpad(vector<byte>& in_out) const
+size_t crypto::padding::X932Padding::Unpad(vector<byte>& in_out) const
 {
 	vector<byte>& in = in_out;
 	vector<byte>& out = in_out;
@@ -52,7 +52,7 @@ int crypto::padding::X932Padding::Unpad(vector<byte>& in_out) const
 	return start;
 }
 
-int crypto::padding::X932Padding::GetPadLength(const size_t len) const
+size_t crypto::padding::X932Padding::GetPadLength(const size_t len) const
 {
 	return block_size_ - len % block_size_ - 1;
 }
